producer/requirements.txt
kafka-python==2.0.3
requests==2.31.0
tweepy==4.14.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0